fiber/requirements.txt
celery==4.3
requests==2.22.0
sanic==19.9.0
sanic-cors==0.9.9.post3
python-socketio==4.3
urllib3==1.25.3
uvloop==0.13.0
